Subcommands for run, invoke, and version
